fix: pom, exception, courtSurface

+++ src/main/java/com/inqool/tennisclub/service/CourtSurfaceService.java
@@ -4,11 +4,12 @@ import com.inqool.tennisclub.data.model.CourtSurfaceEntity;
 import com.inqool.tennisclub.data.repository.CourtSurfaceRepository;
 import com.inqool.tennisclub.exceptions.EntityNotFoundException;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional
 public class CourtSurfaceService {
 
     private final CourtSurfaceRepository courtSurfaceRepository;
@@ -22,12 +23,14 @@ public class CourtSurfaceService {
         return courtSurfaceRepository.save(entity);
     }
 
-    public Optional<CourtSurfaceEntity> findById(Long id) {
-        return Optional.ofNullable(courtSurfaceRepository
+    @Transactional(readOnly = true)
+    public CourtSurfaceEntity findById(Long id) {
+        return courtSurfaceRepository
                 .findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("CourtSurface with id " + id + " not found")));
+                .orElseThrow(() -> new EntityNotFoundException("CourtSurface with id " + id + " not found"));
     }
 
+    @Transactional(readOnly = true)
     public List<CourtSurfaceEntity> findAll() {
         return courtSurfaceRepository.findAll();
     }
@@ -42,9 +45,9 @@ public class CourtSurfaceService {
     }
 
     public void deleteById(Long id) {
-        courtSurfaceRepository
-                .findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("CourtSurface with id " + id + " not found"));
+        if (!courtSurfaceRepository.existsById(id)) {
+            throw new EntityNotFoundException("CourtSurface with id " + id + " not found");
+        }
 
         courtSurfaceRepository.deleteById(id);
     }
